+++ Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v4.2.0
+
+- Upgraded registry to September '20 release 88 (added Libya (LY))
+- Added `IbanBuilder` and `BbanBuilder` types.
+
 ## v4.1.0
 
 ### Changes